backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# AI Services
assemblyai==0.17.0
google-generativeai==0.3.1

# Email & Calendar
sendgrid==6.10.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Background Tasks
celery==5.3.4
redis==5.0.1


httpx==0.25.2

# Google Sheets (for CRM)
gspread==5.12.0
oauth2client==4.1.3

# Audio Processing
pydub==0.25.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1